main/django-react-sheetapp/backend/requirements.txt
# Django Core and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15

# PostgreSQL Database Adapter
psycopg2-binary>=2.9,<3.0

# CORS Handling
django-cors-headers>=4.0,<4.1

# Environment Variable Management
python-dotenv>=1.0,<1.1

# Add other dependencies here as needed, e.g.:
# gunicorn # For production WSGI server
# djoser or dj-rest-auth # For more advanced authentication endpoints
# django-filter # For easy API filtering
# celery # For background tasks
# redis # If using Redis with Celery or for caching